services/result_aggregator_service/pyproject.toml
[project]
name = "result-aggregator-service"
version = "1.0.0"
description = "Result Aggregator Service for HuleEdu Platform"
authors = [
    { name = "HuleEdu Team", email = "team@huleedu.com" },
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "quart",
    "quart-cors",
    "pydantic",
    "pydantic-settings",
    "sqlalchemy",
    "asyncpg",
    "aiohttp",
    "aiokafka",
    "redis",
    "prometheus-client",
    "dishka",
    "python-dotenv",
    "common-core @ file:///${PROJECT_ROOT}/common_core",
    "huleedu-service-libs @ file:///${PROJECT_ROOT}/services/libs",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-mock",
    "pytest-cov",
    "httpx",
    "respx",
]

[tool.pdm]
distribution = false

[tool.pdm.scripts]
dev = "python -m services.result_aggregator_service.app"
test = "pytest tests/"
test-unit = "pytest tests/unit/"
test-integration = "pytest tests/integration/"
format = "ruff format ."
lint = "ruff check ."
typecheck = "mypy ."

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"